% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filename.R
\name{as.filename}
\alias{as.filename}
\alias{as.filename.filename}
\alias{as.filename.character}
\title{Coerce to a filename}
\usage{
as.filename(x, ...)

\method{as.filename}{filename}(x, ...)

\method{as.filename}{character}(x, tag.char = NULL, ...)
}
\arguments{
\item{x}{a \code{character} or a \code{filename}}

\item{...}{other arguments}

\item{tag.char}{character to delimit tags, defaults to \code{'_'}}
}
\value{
a \code{filename} object
}
\description{
This function coerces an object into a \code{filename}, if possible.
}
\examples{
fn <- as.filename("data_raw_2011-01-01.txt")
str(fn)

}
