% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_to_file.R
\name{export_to_file}
\alias{export_to_file}
\title{Export text to a file with a specialized file name}
\usage{
export_to_file(txt, n = 3)
}
\arguments{
\item{txt}{A string}

\item{n}{A number indicating the number of words to use for the file name}
}
\value{
returns TRUE invisibly if it suceeds without error
}
\description{
Export text to a file with a specialized file name
}
