% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filename_B02.R
\name{filename_B02}
\alias{filename_B02}
\title{filename_B02}
\usage{
filename_B02(files = files)
}
\arguments{
\item{files}{List of files. Defaults to files.}
}
\description{
Replaces all multiwell plate 'well' labels in the form 'B2' to the
double-digit form 'B02'. This enables correct ordering of filenames which
include numbers over 10. (Though file systems won't always have a problem,
files in a folder are ordered alphabetically as strings, so B2 is placed
after B10: using B02 solves this problem). Enables correct use of other
functions, such as filename_extension and filename_replicatenumber.
}
\examples{
files <- list.files(path = ".")
filename_B02(files=files)
}
