% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filename_extension.R
\name{filename_extension}
\alias{filename_extension}
\title{filename_extension}
\usage{
filename_extension(filepath = ".", length_with_extension = a,
  length_noextension = b)
}
\arguments{
\item{filepath}{Path to files. Defaults to current directory.}

\item{length_with_extension}{Number of parts to the basename of a file WITH an extension after splitting by "_". Defaults to a.}

\item{length_noextension}{Number of parts to the basename of a file WITHOUT an extension after splitting by "_". Defaults to b.}
}
\description{
Iterates through the provided list of files (alphabetically). If a filename
has part length length_with_extension (as defined by strsplit using "_"), it
is assigned as having an extension/appendix. This function takes the
identified extension/appendix part of the filename and appends it to
following filenames identified as missing an extension/appendix. Designed for
triplicate files in which first file of each group is comprehensively
labelled and the others are not, but should work with any number of
replicates, provided the filepath contains them in the correct order.
}
\examples{
folder <- "FCS_trimmed"
a <- length( unlist(strsplit(basename(list.files(folder)[1]), "_")) )
b <- length( unlist(strsplit(basename(list.files(folder)[2]), "_")) )
filename_extension(filepath = folder, length_with_extension = a, length_noextension = b)
}
