% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filename_prefix.R
\name{filename_prefix}
\alias{filename_prefix}
\title{filename_prefix}
\usage{
filename_prefix(files = files, prefix_n = 3)
}
\arguments{
\item{files}{List of files. Defaults to files.}

\item{prefix_n}{Length of prefix, measured in parts after splitting by "_". Defaults to 3.}
}
\description{
Removes prefixes of specified size in filenames.
}
\examples{
files <- list.files(path = ".")
filename_prefix(files=files, prefix_n = 4)
}
