% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filename_replace.R
\name{filename_replace}
\alias{filename_replace}
\title{filename_replace}
\usage{
filename_replace(files = files, from, to)
}
\arguments{
\item{files}{List of files. Defaults to files.}

\item{from}{String to find.}

\item{to}{String to replace it with.}
}
\description{
Replace a string in a filename with another string.
}
\examples{
files <- list.files(path = ".")
filename_replace(files=files, from="iptg", to="IPTG")
filename_replace(files=files, from="_something_", to="_")
}
