% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filename_replicatenumber.R
\name{filename_replicatenumber}
\alias{filename_replicatenumber}
\title{filename_replicatenumber}
\usage{
filename_replicatenumber(filepath = ".")
}
\arguments{
\item{filepath}{Path to files. Defaults to current directory.}
}
\description{
Append index to end of filename (before the filetype extension such as .fcs)
to show its replicate number, eg. "conditionX_1.fcs", "conditionX_2.fcs".
Assumes triplicate data that is ordered: replicate number appended depends
entirely on each file's position in files (1st = _1, 2nd = _2, 3rd = _3, 4th
= _1, and so on).
}
\examples{
folder <- "FCS_trimmed"
filename_replicatenumber(filepath = folder)
}
