% Generated by roxygen2 (4.0.2): do not edit by hand
\name{is.filename}
\alias{is.filename}
\title{Type checking for filename}
\usage{
is.filename(x)
}
\arguments{
\item{x}{object to check}
}
\value{
a \code{logical} value
}
\description{
This function returns \code{TRUE} if its argument is a \code{filename}
and \code{FALSE} otherwise.
}

