% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/set.R
\name{set_fdate}
\alias{set_fdate}
\title{Set date stamp in a file name}
\usage{
set_fdate(x, date)
}
\arguments{
\item{x}{a \code{character} or a \code{filename}}

\item{date}{new date stamp (\code{character} or \code{Date})}
}
\value{
modified object of the original type
}
\description{
This function sets the date stamp in a file name.
}
\examples{
x <- "data_norm_2011-01-03.txt"
print(set_fdate(x, "2011-01-05"))
}

