% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R
\name{set_fext}
\alias{set_fext}
\title{Set file extension}
\usage{
set_fext(x, ext, all)
}
\arguments{
\item{x}{a \code{character} or a \code{filename}}

\item{ext}{new file extension}

\item{all}{replace the entire extension}
}
\value{
modified object of the original type
}
\description{
This function sets the extension in a file name.
}
\examples{
x <- "data_norm_2011-01-03.txt"
print(set_fext(x, "csv"))

}
