% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R
\name{set_ftime}
\alias{set_ftime}
\title{Set time stamp in a file name}
\usage{
set_ftime(x, time)
}
\arguments{
\item{x}{a \code{character} or a \code{filename}}

\item{time}{new time stamp (\code{character} or \code{POSIXct})}
}
\value{
modified object of the original type
}
\description{
This function sets the time stamp in a file name.
}
\examples{
x <- "data_norm_20110103T093015.txt"
# change the time to 30 seconds past 2:45 p.m. 
print(set_ftime(x, "144530"))
# to change the date, time must be specified as well
print(set_ftime(x, "20110505T101500"))

}

