% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tag.R
\name{tag}
\alias{tag}
\title{Insert tag or extension and coerce to character}
\usage{
tag(x, ...)
}
\arguments{
\item{x}{a \code{filename} or \code{character}}

\item{...}{arguments passed to \code{\link{insert}}}
}
\value{
a \code{character} vector
}
\description{
This fucntion inserts a tag or extension into a file name and returns
a \code{charcter} vector.
}
\examples{
x <- "data.txt"
y <- tag(x, "qc")
print(y)
f <- as.filename(x)
g <- tag(f, "qc")
print(g)
}

