% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim.R
\name{trim_ext}
\alias{trim_ext}
\title{Trim extensions from a file name}
\usage{
trim_ext(x, n)
}
\arguments{
\item{x}{a \code{character} or a \code{filename}}

\item{n}{number of extensions to trim off the end}
}
\value{
modified object of the original type
}
\description{
This function trims extensions from a file name.
}
\examples{
x <- "path/data.txt.gz"
print(trim_ext(x))

}
