% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filename.R
\name{as.character.filename}
\alias{as.character.filename}
\title{Coerce a character to a filename}
\usage{
\method{as.character}{filename}(x, tag.char = NULL, simplify = FALSE, ...)
}
\arguments{
\item{x}{a \code{filename} object}

\item{tag.char}{character to delimit tags, defaults to \code{'_'}}

\item{simplify}{if \code{TRUE}, all timestamps are omitted}

\item{...}{unused arguments}
}
\value{
a \code{character} vector
}
\description{
This function coerces a \code{filename} into a character.
}
\examples{
x <- "data_post_2011-01-02.txt"
fn <- as.filename(x)
print(as.character(fn))

}
