% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filename_replace.R
\name{filename_replace}
\alias{filename_replace}
\title{filename_replace}
\usage{
filename_replace(filepath = ".", from, to)
}
\arguments{
\item{filepath}{Path to files. Defaults to current directory.}

\item{from}{String to find.}

\item{to}{String to replace it with.}
}
\description{
Replace a string in a filename with another string.
}
\examples{
folder <- "FCS_trimmed"
filename_replace(filepath = folder, from="iptg", to="IPTG")
filename_replace(filepath = folder, from="100ngmlatc", to="100")
filename_replace(filepath = folder, from="_something_", to="_")
}
