% Generated by roxygen2 (4.0.2): do not edit by hand
\name{set_fext}
\alias{set_fext}
\title{Set file extension}
\usage{
set_fext(x, ext)
}
\arguments{
\item{x}{a \code{character} or a \code{filename}}

\item{ext}{new file extension}
}
\value{
modified object of the original type
}
\description{
This function sets the extension in a file name.
}
\examples{
x <- "data_norm_2011-01-03.txt"
print(set_fext(x, "csv"))
}

